ai_service/requirements.txt
ultralytics==8.1.2
opencv-python==4.9.0.80
torch==2.2.0
numpy==1.26.4
psutil==5.9.8
python-dotenv==1.0.1
websockets==12.0
requests==2.31.0
python-json-logger==2.0.7
pygame==2.5.2 SSS